refactor: move shared store code to shared folder
